requirements.txt
google-cloud-speech
google-cloud-storage
gradio
Wave
websockets
